+++ src/arch/x86.rs
@@ -7,4 +7,5 @@ pub mod idt;
 pub mod interrupts;
 pub mod kernel_mutex;
 pub mod scheduler;
+pub mod syscall;
 pub mod vmm;
